requirements.txt
requests==2.11.1
win-unicode-console==0.5
CacheControl==0.11.7
lockfile==0.12.2
python-dotenv==0.6.0